@@ -1,15 +1,15 @@
 {
-    "name": "lingxi/ali-opensearch-sdk",
+    "name": "lianran111/ali-opensearch-sdk",
     "autoload": {
         "psr-4": {
             "Lingxi\\AliOpenSearch\\": "src/"
         }
     },
     "require": {
-        "php": "^5.5.9 || ^7.0",
+        "php": "^7.2",
         "laravel/scout": "~2.0 || ~3.0",
-        "illuminate/events": "~5.1",
-        "illuminate/contracts": "~5.1"
+        "illuminate/events": "^5.1",
+        "illuminate/contracts": "^5.1"
     },
     "require-dev": {
         "phpunit/phpunit": "^4.8 || ^5.0",
