@@ -4,25 +4,25 @@ namespace Lingxi\AliOpenSearch\Events;
 
 class DocSyncEvent
 {
-	public $indexName;
+    public $indexName;
 
-	public $tableName;
+    public $tableName;
 
-	public $data;
+    public $data;
 
-	public $type;
+    public $type;
 
-	public $success;
+    public $success;
 
-	public $message;
+    public $message;
 
-	public function __construct($indexName, $tableName, array $data, $type, $success = true, $message = '')
-	{
-		$this->data = $data;
-		$this->type = $type;
-		$this->success = $success;
-		$this->message = $message;
-		$this->indexName = $indexName;
-		$this->tableName = $tableName;
-	}
+    public function __construct($indexName, $tableName, array $data, $type, $success = true, $message = '')
+    {
+        $this->data      = $data;
+        $this->type      = $type;
+        $this->success   = $success;
+        $this->message   = $message;
+        $this->indexName = $indexName;
+        $this->tableName = $tableName;
+    }
 }
